+++ src/PowerShell/Microsoft.WinGet.Client.Engine/Common/Constants.cs
@@ -1,0 +1,19 @@
+﻿// -----------------------------------------------------------------------------
+// <copyright file="Constants.cs" company="Microsoft Corporation">
+//     Copyright (c) Microsoft Corporation. Licensed under the MIT License.
+// </copyright>
+// -----------------------------------------------------------------------------
+
+namespace Microsoft.WinGet.Client.Engine.Common
+{
+    /// <summary>
+    /// This class contains all of the configurable constants for this project.
+    /// </summary>
+    internal static class Constants
+    {
+        /// <summary>
+        /// Name of PATH environment variable.
+        /// </summary>
+        public const string PathEnvVar = "PATH";
+    }
+}
