+++ src/PowerShell/Microsoft.WinGet.Client.Engine/Common/IntegrityCategory.cs
@@ -4,7 +4,7 @@
 // </copyright>
 // -----------------------------------------------------------------------------
 
-namespace Microsoft.WinGet.Client.Common
+namespace Microsoft.WinGet.Client.Engine.Common
 {
     /// <summary>
     /// The type of the integrity check failure.
