+++ src/PowerShell/Microsoft.WinGet.Client.Engine/Exceptions/InvalidSourceException.cs
@@ -4,10 +4,10 @@
 // </copyright>
 // -----------------------------------------------------------------------------
 
-namespace Microsoft.WinGet.Client.Exceptions
+namespace Microsoft.WinGet.Client.Engine.Exceptions
 {
     using System;
-    using Microsoft.WinGet.Client.Properties;
+    using Microsoft.WinGet.Client.Engine.Properties;
 
     /// <summary>
     /// Invalid source.
