+++ src/PowerShell/Microsoft.WinGet.Client.Engine/Properties/Resources.resx
